Fix crash in active_highlights debug assertion
It's possible for active_highlights to be empty, in which case the debug
assertion crashes, since layer_start is 1. The code to calculate
`parent_start` in `advance` already solved this by restricting the index
to the length of the `active_highlights` array, but the debug assertion
did not.
Move this code into a new method `parent_start` and use it in both cases.