support using `; extends` in a new read_query_extends fn
makes a new function `read_query_extends`, that is similar to
`read_query`, but it takes a closure that returns an Iterator of Strings
in order of priority, and uses that to handle `; extends` directives,
going down one level of priority for each directive.

the `; extends` directive makes it, so that you can define custom queries
without having to copy all of the existing ones, by just doing
`; extends` in your file and then writing your own custom queries.